feat: add example route
