+++ src/pages/App.js
@@ -2,6 +2,7 @@ import {Suspense} from 'react'
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom'
 import Shuttle from '../pages/Shuttle'
 import TokenList from '../pages/Shuttle/TokenList'
+import Example from './Example'
 
 function App() {
   return (
@@ -15,6 +16,10 @@ function App() {
             {/* TODO: Remove */}
             <Route path="/tokenlist">
               <TokenList />
+            </Route>
+            {/* TODO: Remove */}
+            <Route path="/example">
+              <Example />
             </Route>
             <Route path="/">
               <Shuttle />
